refactor package support
